Fade Scene
Added FadeScene to bottom of UI Prefab.  Had a Fade Time in second that can be adjusted easily for all scenes.  Runs the Fade Out from the Change Scene Script, and has been attached to all the Change Scenes that I have been able to find.  If it is missing from one, it won't throw errors, but will rather just change scene directly.