Fix for Issue 77 and 61
Added a Quest Icon that can easily be added to Quest Givers.  Will give a more in depth way of adding during meeting.

While not perfect, have made fixes to the NPCOneLiner.cs script to allow people to click on them and see the dialogue.  Had to make some edits to the TownAfternoon Front Gate Edges, since they prevented the player from talking to guards.  Haven't tested for every character so will have to be on the look out for similar issues.

Changed Items with a unique boolean, to prevent the game from repopulating unique items after initial pickup.

Refactors Scythe and PianoKey Checks.  Might change that again and just make them unique.