Quest Icon changes
Added a complete VAR in Ink for piano_quest.  Every quest will need 3 VAR, one for when it is ready to be give, one for when player has it, and one for when it is complete.

It's not the greatest code because it needs to grab info and make an observer for the ink.  I may come back to this in the future but it works for now.