Cutscene Issue fixed
Not a particularly elegant fix, but it works.  Basically had a Update on NPCOneLiner that closed dialogbox on right mouse click, as that is currently only way to get out of it.  Had to add a boolean to tell if it was a OneLiner open or not, before attempting close.  Will look at better way to do this in the near future.