Quest  and Save System v0.1
Still needs work and refactoring.  Some design choices I need to look back on and change.  It's rather comprehensive and currently not easy to add new quests, so I am going to look for a way to either make that part easier, or make it more dynamic.

May break other aspects of the game, if it does, please let me know and I will find a way to fix them.