First Refactoring
Changes to various scripts with the idea of reducing or making it easier to work with code.

Got rid of PlaySoundQ and made a simple PlaySound with a single function.  Then add a function event in the animator for the Scythe Attacks, to play the sound when it starts swing.  This PlaySound function can easily be added to other animations that need sound at specific times.  PlaySoundQ was working on constantly updates when it was unneeded to.