More refactoring
Did a little bit or reorganizing as well but more to do on that later.  Got the Music Player stuff outside of the GameManager.  It is finally looking trim, though can still a bit more work on it.  Seeing a lot of small little bugs in the game, some I am sure are due to my changes.  Not quite ready to push yet.