Final Changes for this Refactoring
Deleted some instances where a script did not exist anymore from GameObjects.  Added a script to change the player's outfit, though will be finding a better way to do that in the future.

Changes to the Inventory are not done, as they are not clickable and I don't believe there are any ways to open up the book at the moment.  However I will be fixing those next week along with knocking out a bunch of the other issues before I get back to Refactoring.