Bug Fixings
Fixed issue with Quest not showing up in Log after getting.  Along with that, I changed the VAR names a bit for the other quests and hope that we can continue to use that naming format to make it easier to parse through them.

Naming has_(quest name),  ready_for_(quest name), and completed_(quest name).  We can change it but I would like to have it consistent in it's naming to make it easier to make the scripts more dynamic.

Fixed bug where Trivia didn't pop up the first time it was opened.  Trivia is also on the list of things that need to be refactored.

Lastly did a slight change to the windchimes to make it where you won't hear them all over the FramMorning map.  Fairly easy change in Unity.  In Audio Source, there is a setting called 3D Sound Settings.  Expand that and you will see a graph.  The red volume is how far out the camera will pick up the sound, and how loud it will be at that distance.  You can change the max distance along with how long it will be easily with this.