Audio playing in every dialogue issue.
In Audio Source these is a setting called Play on Awake.  Originally the audio clip isn't set, so it doesn't play anything when dialogue opens, but when you get the quest, it sets the audio to the quest play.  After that, every time the dialogue is turned on, it plays the clip.  Simply turning this setting off, prevents it from playing until specifically told to in scripts.