Rudimentary Settings Menu with Volume Controller
Created a simple settings menu and volume controller.  Needs design work (you can do a lot with the slider or create a custom one).  Volume control works and saves in PlayerPrefs which is good for overall game settings such as this.  Will carry from scene to scene and even on new loads.  Not sure how much we want a save/load in here, so will discuss that more in depth on Saturday.