Added Test Quests
On Title Screen there is now a button for Test Quest Menu, and the two current quests can be started right from there.  This should make testing for new quests significantly faster, without having to mess with the ink file directly.  This button is set to be destroyed if it is not in the Editor, so it shouldn't effect builds, but I have not tested that, just used an example that should work.

I have set it up for the Piano and Trombone quest, but if I missed any variables for these quests, let me know and I can change them in the test script.

Along with that, I fixed the issue where new games start on 0 volume, but if you game has already been loaded and the setting hasn't changed, it will still be stuck at 0 till you change it in the settings menu.