General Bugfixing and Organizing - More in Dsecription.
Lot of 'changes' but most of them are just reorganizing.  Took all the loose scripts in the top level of Scripts, and sorted them a bit.  Anything that didn't have a direct location I put under a new folder called '_System' that will show up as the first folder in the file structure.

Added function to turn off Interact Light.  Basically in the InteractionAnimation script, I have a new variable that will shut off the light if it is set.  Though upon reentering a scene, it will show up as the same most likely.  Will do some more work on that in the future.

Made the Scarecrows collider bigger and made it higher in the order in hopes that it is above the interactable wheat colliders.

Fixed bug where if you clicked on something that was a different icon and then clicked back on the ground, it will still show the other icon.  now it will properly display the normal cursor when you click on the ground in general.